chore(components): migrate aiplatform.model and aiplatform.endpoint components to v1 namespace
PiperOrigin-RevId: 526757962
