+++ components/google-cloud/google_cloud_pipeline_components/v1/endpoint/__init__.py
@@ -14,14 +14,22 @@
 """Google Cloud Pipeline Endpoint components."""
 
 import os
-
+from kfp import components
 from .create_endpoint import component as create_endpoint_component
 from .deploy_model import component as deploy_model_component
 
 __all__ = [
     'EndpointCreateOp',
+    'EndpointDeleteOp',
     'ModelDeployOp',
+    'ModelUndeployOp',
 ]
 
 EndpointCreateOp = create_endpoint_component.endpoint_create
+EndpointDeleteOp = components.load_component_from_file(
+    os.path.join(os.path.dirname(__file__), 'delete_endpoint/component.yaml')
+)
 ModelDeployOp = deploy_model_component.model_deploy
+ModelUndeployOp = components.load_component_from_file(
+    os.path.join(os.path.dirname(__file__), 'undeploy_model/component.yaml')
+)
