+++ components/google-cloud/google_cloud_pipeline_components/v1/endpoint/delete_endpoint/__init__.py
@@ -11,21 +11,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Google Cloud Pipeline Model components."""
-
-import os
-from kfp import components
-from .export_model import component as model_export_component
-from .upload_model import component as model_upload_component
-
-__all__ = [
-    'ModelExportOp',
-    'ModelUploadOp',
-    'ModelDeleteOp',
-]
-
-ModelExportOp = model_export_component.model_export
-ModelUploadOp = model_upload_component.model_upload
-ModelDeleteOp = components.load_component_from_file(
-    os.path.join(os.path.dirname(__file__), 'delete_model/component.yaml')
-)
+"""Delete Endpoint Component."""
