+++ components/google-cloud/google_cloud_pipeline_components/v1/endpoint/delete_endpoint/component.yaml
