+++ components/google-cloud/google_cloud_pipeline_components/v1/model/delete_model/component.yaml
