+++ components/google-cloud/tests/v1/testdata/create_and_delete_endpoint_pipeline.json
